+++ CHANGELOG.md
@@ -16,3 +16,4 @@ _`**VERSIONE 1.0**`_
 - Aggiunto `ModelloEbay`
 - Aggiunto `Test Colore search button`
 - Aggiunto `Test posizione su google Maps e su demons`
+- Aggiunto `mobile properties`